Substitute QueryDsl and CountBuilder by SingleQueryBuilder so it could be used as a query registry to construct single queries accepted by Count and DeleteByQuerey commands.
