--- src/PlainElastic.Net/Builders/Queries/QueryDsl.cs
@@ -1,12 +1,0 @@
-﻿using PlainElastic.Net.Utils;
-
-namespace PlainElastic.Net.Queries
-{
-    public class QueryDsl<T> : Query<T>
-    {
-        protected override string ApplyJsonTemplate(string body)
-        {
-            return "{0}".AltQuoteF(body);            
-        }
-    }
-}
